cart and order in BL

+++ BL/BO/OrderForList.cs
@@ -7,7 +7,7 @@ public class OrderForList
     public int OrderId { get; set; }
     public Enums.State state { get; set; }
     public int Amount { get; set; }
-    public int TotalPrice { get; set; }
+    public Double TotalPrice { get; set; }
 
     public override string ToString() => $@"
 order id: ={OrderId}
